Add Atmosphere class
This is the second class from the uvagoa API added to the repository. It
allows to store information regarding the atmospheric properties, e.g.
the atmospheric pressure, the vertical ozone content, the total amount
of water vapour, the Angstrom parameters, the single scattering albedo
and the asymmetry parameter. It also allows the access to the absorption
coefficients in read-only mode.