Rename Geometry nscen property to ngeo
Both Geometry and Atmosphere had a nscen property to know the actual
size of the instance, but this could lead to confusion because they
were named the same. For this reason, the Geometry size is renamed
from nscen to ngeo, and the docstrings mention it now as the number
of geometries within the Geometry instance (instance of number the
of scenarios).