Remove all squeeze options from solo
All the references to the squeeze of axes in the outputs of any method
have been removed, because they do not give any particular benefit and
they may cause problems to the final user due to a bad matching of axes.
Currently, all the methods work as expected, but there should be a
further inspection to check that all outputs have the expected shapes.

This commits adds the proposal from issue #5.