Add option to run radtran as executable
As a new feature, this commits adds the possibility to run radtran
directly as en executable, by typing in the console:

    python3 radtran.py --geo=geopath.dat --atm=atmpath.dat
    --out=outpath.dat (--no-coupling)

where 'geopath.dat', 'atmpath.dat' and 'outpath.dat' denote the paths to
the Geometry instance, the Atmosphere instance and the output location,
so  that if 'outpath.dat' is given, the irradiances will be written to
'outpath_glb.dat', 'outpath_dir.dat' and 'outpath_dif.dat'. If the flag
--no-coupling is given, then the Rayleigh-aerosol coupling is not used.

This commits adds the request from issue #6.